Consolidated workspace logic
